worker/requirements.txt
httpx[http2]==0.27.0
redis==5.0.1
psycopg2-binary==2.9.9
selectolax==0.3.21
tldextract==5.1.1
python-dotenv==1.0.0
anthropic==0.18.1
pydantic==2.6.1
